requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
cricapi==1.2
dateparser==1.0.0
discord.py==1.6.0
idna==2.10
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
yarl==1.6.3
